изменен класс Game

+++ snakesfml/Picture.cpp
@@ -1,5 +1,5 @@
 #include "Picture.h"
-Picture::Picture(std::string pic_name, float x_position, float y_position, int x1_im, int y1_im, int x2_im, int y2_im) : x_position(x_position), y_position(y_position),
+Picture::Picture(std::string pic_name, int x_position, int y_position, int x1_im, int y1_im, int x2_im, int y2_im) : x_position(x_position), y_position(y_position),
 x1_im(x1_im), x2_im(x2_im), y1_im(y1_im), y2_im(y2_im), pic_name(pic_name)
 {
 	image.loadFromFile("pictures/" + pic_name);
