+++ snakesfml/Picture.h
@@ -13,6 +13,6 @@ public:
 	sf::Texture texture;
 	sf::Sprite sprite;
 	std::string pic_name; //name of picture
-	Picture(std::string pic_name, float x_position = 0, float y_position = 0, int x1_im = -1, int y1_im = -1, int x2_im = -1, int y2_im = -1);
+	Picture(std::string pic_name, int x_position = 0, int y_position = 0, int x1_im = -1, int y1_im = -1, int x2_im = -1, int y2_im = -1);
 	void Create_Mask(int r, int g, int b, int a);
 };
